Re-writing Mstate-funcall and helpers
Beginning of re-write of Mstate-funcall, or more specifically, getFunctionExecutionEnvironment and its related functions. This rewrite re-orders the creation of a function environment to more closely match the guidelines in the assignment: Create the environment using the function closure, add an empty layer of scope, and then bind the actual parameters to the formal parameters.

This commit also includes a number of re-named functions, with a few more to come. Re-names are intended to make each function's purpose more clear; for example, getEnvironment is renamed to initializeFunctionExecutionEnvironment.